update ui using taliwind & daisyui

+++ routes/web.php
@@ -20,17 +20,17 @@ use App\Http\Controllers\HomeController;
 
 Route::get('/', [HomeController::class, 'index']);
 
-Route::get('/admin', [TaskController::class, 'index'])->middleware('auth');
+Route::get('/admin', [TaskController::class, 'index'])->middleware('is_admin');
 
-Route::get('/admin/create', [TaskController::class, 'create']);
+Route::get('/admin/create', [TaskController::class, 'create'])->middleware('is_admin');
 
 Route::get('/admin/{id}', [TaskController::class, 'show']);
 
 Route::post('/admin', [TaskController::class, 'store']);
 
-route::get('/admin/{id}/edit', [TaskController::class, 'edit']);
+route::get('/admin/{id}/edit', [TaskController::class, 'edit'])->middleware('is_admin');
 
-Route::patch('/admin/{id}', [TaskController::class, 'update']);
+Route::patch('/admin/{id}', [TaskController::class, 'update'])->middleware('is_admin');
 
-Route::delete('/admin/{id}', [TaskController::class, 'delete']);
+Route::delete('/admin/{id}', [TaskController::class, 'delete'])->middleware('is_admin');
